chore(pg/searchplayer): this function should only gives players without sub objects
INTRODUCE BREAKING CHANGE.
We want to modify all the business logic to usecase layer instead of there. We want to minimize all tasks on backend an controller layers. This functions breaks a lot of usecases as it returns objects with less information than before.

It requires a second commit to move logic to usecase layer.